Split node type definitions to a separate crate
