+++ advanca-node-primitives/Cargo.toml
@@ -1,0 +1,54 @@
+[package]
+authors = ['Advanca Authors']
+edition = '2018'
+license = 'GPL-3.0'
+name = 'advanca-node-primitives'
+version = '0.3.0'
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies.frame-support]
+default-features = false
+tag = 'v2.0.0-rc5+2'
+git = 'https://github.com/paritytech/substrate.git'
+version = '2.0.0-rc5'
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.3.1'
+
+[dependencies.smart-default]
+default-features = false
+version = "0.6.0"
+
+[dependencies.frame-system]
+default-features = false
+tag = 'v2.0.0-rc5+2'
+git = 'https://github.com/paritytech/substrate.git'
+version = '2.0.0-rc5'
+
+[dependencies.sp-core]
+default-features = false
+tag = 'v2.0.0-rc5+2'
+git = 'https://github.com/paritytech/substrate.git'
+version = '2.0.0-rc5'
+
+[dependencies.sp-runtime]
+default-features = false
+tag = 'v2.0.0-rc5+2'
+git = 'https://github.com/paritytech/substrate.git'
+version = '2.0.0-rc5'
+
+[features]
+default = ["std"]
+std = [
+	"codec/std",
+	"frame-system/std",
+	"sp-core/std",
+	"sp-runtime/std",
+]
