+++ pallets/advanca-core/Cargo.toml
@@ -5,6 +5,9 @@ license = 'GPL-3.0'
 name = 'advanca-core'
 version = '0.3.0'
 
+[dependencies]
+advanca-node-primitives = { path = "../../advanca-node-primitives/" }
+
 [package.metadata.docs.rs]
 targets = ['x86_64-unknown-linux-gnu']
